Persistent RTC-backed stopwatch + 99-day range + countdown elapsed alert
- Stopwatch: persist elapsed/start-epoch/view-active to NVS, restore
  on boot with RTC epoch catch-up across power loss
- Stopwatch: widen internal type from uint32_t → uint64_t for 99-day range
- Stopwatch display: extend SS.cc tier from 60s → 99s for consistency
- Countdown alert: show elapsed time since expiry (capped at 99:59)
  instead of blinking 00:00
- Add NVS keys sw_elapsed (uint64), sw_start_ep (int32), sw_view (bool)

+++ src/TimerRenderer.h
@@ -27,7 +27,7 @@ public:
 
 private:
     void drawTimerDuration(uint32_t totalSeconds, bool blinkSeparator);
-    void drawStopwatchTime(uint32_t totalSec, uint8_t centisec, bool running);
+    void drawStopwatchTime(uint64_t totalSec, uint8_t centisec, bool running);
     void renderStopwatch();
     void renderCountdown();
     void renderCountdownAlert();
